[WIP] Separated segment handling logic from IndexedErgoAddress
