+++ src/main/scala/org/ergoplatform/nodeView/history/extra/IndexedErgoAddress.scala
@@ -1,183 +1,33 @@
 package org.ergoplatform.nodeView.history.extra
 
 import org.ergoplatform.ErgoAddressEncoder
-import org.ergoplatform.http.api.SortDirection.{ASC, DESC, Direction}
-import org.ergoplatform.nodeView.history.{ErgoHistory, ErgoHistoryReader}
+import org.ergoplatform.nodeView.history.ErgoHistoryReader
 import org.ergoplatform.nodeView.history.extra.ExtraIndexer.{ExtraIndexTypeId, fastIdToBytes}
-import org.ergoplatform.nodeView.history.extra.IndexedErgoAddress.{getBoxes, getFromSegments, getTxs}
-import org.ergoplatform.nodeView.history.extra.IndexedErgoAddressSerializer.{boxSegmentId, txSegmentId}
 import org.ergoplatform.settings.Algos
 import scorex.core.serialization.ErgoSerializer
-import scorex.util.{ModifierId, ScorexLogging, bytesToId}
+import scorex.util.{ModifierId, bytesToId}
 import scorex.util.serialization.{Reader, Writer}
 import sigmastate.Values.ErgoTree
-import spire.ClassTag
-
-import scala.collection.mutable.ArrayBuffer
-import spire.syntax.all.cfor
-
-import java.lang.Math.abs
 
 /**
   * An index of an address (ErgoTree)
   * @param treeHash    - hash of the corresponding ErgoTree
   * @param txs         - list of numberic transaction indexes associated with this address
   * @param boxes       - list of numberic box indexes associated with this address, negative values indicate the box is spent
-  * @param balanceInfo - balance information (Optional because fragments do not contain it)
   */
-case class IndexedErgoAddress(treeHash: ModifierId,
-                              txs: ArrayBuffer[Long],
-                              boxes: ArrayBuffer[Long],
-                              balanceInfo: Option[BalanceInfo]) extends ExtraIndex with ScorexLogging {
+case class IndexedErgoAddress(treeHash: ModifierId)
+  extends Segment[IndexedErgoAddress](treeHash, id => IndexedErgoAddress(id)) with ExtraIndex {
 
   override lazy val id: ModifierId = treeHash
-  override def serializedId: Array[Byte] = fastIdToBytes(id)
-
-  // Internal segment buffer used when spending boxes
-  private[extra] val segments: ArrayBuffer[IndexedErgoAddress] = ArrayBuffer.empty[IndexedErgoAddress]
-
-  private[extra] var boxSegmentCount: Int = 0
-  private[extra] var txSegmentCount: Int = 0
+  override def serializedId: Array[Byte] = fastIdToBytes(treeHash)
 
   /**
-    * @return total number of transactions associated with this address
-    */
-  def txCount(implicit segmentTreshold: Int): Long = segmentTreshold * txSegmentCount + txs.length
+   * Balance information (Optional because fragments do not contain it)
+   */
+  var balanceInfo: Option[BalanceInfo] = None
 
-  /**
-    * @return total number of boxes associated with this address
-    */
-  def boxCount(implicit segmentTreshold: Int): Long = segmentTreshold * boxSegmentCount + boxes.length
-
-  /**
-    * Copied from [[java.util.Arrays.binarySearch]]
-    */
-  private def binarySearch(a: ArrayBuffer[Long], key: Long): Int = {
-    var low: Int = 0
-    var high: Int = a.length - 1
-
-    while (low <= high) {
-      val mid = (low + high) >>> 1
-      val midVal = abs(a(mid)) // ignore negativity
-
-      if (midVal < key)
-        low = mid + 1
-      else if (midVal > key)
-        high = mid - 1
-      else
-        return mid // key found
-    }
-    -1 // key not found.
-  }
-
-  /**
-    * Retrieve segment with specified id from buffer or database
-    * @param history - database handle to search, if segment is not found in buffer
-    * @param id - address segment to search for
-    * @return
-    */
-  private def getSegmentFromBufferOrHistroy(history: ErgoHistoryReader, id: ModifierId): Int = {
-    cfor(segments.length - 1)(_ >= 0, _ - 1) { i =>
-      if(segments(i).id.equals(id)) return i
-    }
-    segments += history.typedExtraIndexById[IndexedErgoAddress](id).get
-    segments.length - 1
-  }
-
-  /**
-    * Locate which segment the given box number is in and change its sign, meaning it spends unspent boxes and vice versa.
-    * @param boxNum - box number to locate
-    * @param history - database to retrieve segments from
-    */
-  private[extra] def findAndModBox(boxNum: Long, history: ErgoHistoryReader): Unit = {
-    val boxNumAbs = abs(boxNum)
-    val inCurrent: Int = binarySearch(boxes, boxNumAbs)
-    if(inCurrent >= 0) { // box found in current box array
-      boxes(inCurrent) = -boxes(inCurrent)
-    } else { // box is in another segment, use binary search to locate
-      var n = 0
-      var low = 0
-      var high = boxSegmentCount - 1
-      while(low <= high) {
-        val mid = (low + high) >>> 1
-        n = getSegmentFromBufferOrHistroy(history, boxSegmentId(treeHash, mid))
-        if(abs(segments(n).boxes.head) < boxNumAbs &&
-           abs(segments(n).boxes.last) < boxNumAbs)
-          low = mid + 1
-        else if(abs(segments(n).boxes.head) > boxNumAbs &&
-                abs(segments(n).boxes.last) > boxNumAbs)
-          high = mid - 1
-        else
-          low = high + 1 // break
-      }
-      val i: Int = binarySearch(segments(n).boxes, boxNumAbs)
-      if(i >= 0)
-        segments(n).boxes(i) = -segments(n).boxes(i)
-      else
-        log.warn(s"Box $boxNum not found in any segment of parent address when trying to spend")
-    }
-  }
-
-  /**
-    * Get a range of the transactions associated with this address
-    * @param history - history to use
-    * @param offset  - items to skip from the start
-    * @param limit   - items to retrieve
-    * @return array of transactions with full bodies
-    */
-  def retrieveTxs(history: ErgoHistoryReader, offset: Int, limit: Int)(implicit segmentTreshold: Int): Array[IndexedErgoTransaction] =
-    getFromSegments(history, treeHash, offset, limit, txSegmentCount, txs, txSegmentId, _.txs, getTxs)
-
-  /**
-    * Get a range of the boxes associated with this address
-    * @param history - history to use
-    * @param offset  - items to skip from the start
-    * @param limit   - items to retrieve
-    * @return array of boxes
-    */
-  def retrieveBoxes(history: ErgoHistoryReader, offset: Int, limit: Int)(implicit segmentTreshold: Int): Array[IndexedErgoBox] =
-    getFromSegments(history, treeHash, offset, limit, boxSegmentCount, boxes, boxSegmentId, _.boxes, getBoxes)
-
-  /**
-    * Get a range of the boxes associated with this address that are NOT spent
-    * @param history - history to use
-    * @param offset  - items to skip from the start
-    * @param limit   - items to retrieve
-    * @param sortDir - whether to start retreival from newest box ([[DESC]]) or oldest box ([[ASC]])
-    * @return array of unspent boxes
-    */
-  def retrieveUtxos(history: ErgoHistoryReader, offset: Int, limit: Int, sortDir: Direction): Array[IndexedErgoBox] = {
-    val data: ArrayBuffer[IndexedErgoBox] = ArrayBuffer.empty[IndexedErgoBox]
-    sortDir match {
-      case DESC =>
-        data ++= boxes.filter(_ > 0).map(n => NumericBoxIndex.getBoxByNumber(history, n).get)
-        var segment: Int = boxSegmentCount
-        while (data.length < (limit + offset) && segment > 0) {
-          segment -= 1
-          history.typedExtraIndexById[IndexedErgoAddress](boxSegmentId(treeHash, segment)).get.boxes
-            .filter(_ > 0).map(n => NumericBoxIndex.getBoxByNumber(history, n).get) ++=: data
-        }
-        data.reverse.slice(offset, offset + limit).toArray
-      case ASC =>
-        var segment: Int = 0
-        while (data.length < (limit + offset) && segment < boxSegmentCount) {
-          data ++= history.typedExtraIndexById[IndexedErgoAddress](boxSegmentId(treeHash, segment)).get.boxes
-            .filter(_ > 0).map(n => NumericBoxIndex.getBoxByNumber(history, n).get)
-          segment += 1
-        }
-        if(data.length < (limit + offset))
-          data ++= boxes.filter(_ > 0).map(n => NumericBoxIndex.getBoxByNumber(history, n).get)
-        data.slice(offset, offset + limit).toArray
-    }
-  }
-
-  /**
-    * Associate transaction index with this address
-    * @param tx - numeric transaction index
-    * @return this address
-    */
-  private[extra] def addTx(tx: Long): IndexedErgoAddress = {
-    if(txs.lastOption.getOrElse(-1) != tx) txs += tx // check for duplicates
+  private[extra] def initBalance: IndexedErgoAddress = {
+    balanceInfo = Some(BalanceInfo())
     this
   }
 
@@ -187,9 +37,9 @@ case class IndexedErgoAddress(treeHash: ModifierId,
     * @param record - whether to add box to boxes list, used in rollbacks (true by default)
     * @return this address
     */
-  private[extra] def addBox(iEb: IndexedErgoBox, record: Boolean = true): IndexedErgoAddress = {
+  override private[extra] def addBox(iEb: IndexedErgoBox, record: Boolean = true): IndexedErgoAddress = {
     if(record) boxes += iEb.globalIndex
-    balanceInfo.get.add(iEb.box)
+    balanceInfo.foreach(_.add(iEb.box))
     this
   }
 
@@ -199,98 +49,24 @@ case class IndexedErgoAddress(treeHash: ModifierId,
     * @param historyOpt - history handle to update address fragment if spent box is old
     * @return this address
     */
-  private[extra] def spendBox(iEb: IndexedErgoBox, historyOpt: Option[ErgoHistoryReader] = None)(implicit ae: ErgoAddressEncoder): IndexedErgoAddress = {
-    balanceInfo.get.subtract(iEb.box)
-
-    if(historyOpt.isEmpty)
-      return this
-
-    findAndModBox(iEb.globalIndex, historyOpt.get)
-
+  override private[extra] def spendBox(iEb: IndexedErgoBox, historyOpt: Option[ErgoHistoryReader] = None)(implicit ae: ErgoAddressEncoder): IndexedErgoAddress = {
+    if(historyOpt.isDefined)
+      findAndModBox(iEb.globalIndex, historyOpt.get)
+    balanceInfo.foreach(_.subtract(iEb.box))
     this
   }
 
   /**
-    * Rollback the state of this address and of boxes associted with it
-    * @param txTarget  - remove transaction numbers above this number
-    * @param boxTarget - remove box numbers above this number and revert the balance
-    * @param _history  - history handle to update address in database
-    * @return modifier ids to remove
-    */
-  private[extra] def rollback(txTarget: Long, boxTarget: Long, _history: ErgoHistory)(implicit segmentTreshold: Int): Array[ModifierId] = {
-
-    if((txCount == 0 && boxCount == 0) || // address already rolled back
-       (txs.last <= txTarget && abs(boxes.last) <= boxTarget)) // no rollback needed
-      return Array.empty[ModifierId]
-
-    def history: ErgoHistoryReader = _history.getReader
-
-    val toSave: ArrayBuffer[ExtraIndex] = ArrayBuffer.empty[ExtraIndex]
-    val toRemove: ArrayBuffer[ModifierId] = ArrayBuffer.empty[ModifierId]
-
-    // filter tx numbers
-    do {
-      val tmp = txs.takeWhile(_ <= txTarget)
-      txs.clear()
-      txs ++= tmp
-      if(txs.isEmpty && txSegmentCount > 0) { // entire current tx set removed, retrieving more from database if possible
-        val id = txSegmentId(treeHash, txSegmentCount - 1)
-        txs ++= history.typedExtraIndexById[IndexedErgoAddress](id).get.txs
-        toRemove += id
-        txSegmentCount -= 1
-      }
-    }while(txCount > 0 && txs.last > txTarget)
-
-    // filter box numbers
-    do {
-      val tmp = boxes.takeWhile(abs(_) <= boxTarget)
-      boxes.clear()
-      boxes ++= tmp
-      if(boxes.isEmpty && boxSegmentCount > 0) { // entire current box set removed, retrieving more from database if possible
-        val id = boxSegmentId(treeHash, boxSegmentCount - 1)
-        boxes ++= history.typedExtraIndexById[IndexedErgoAddress](id).get.boxes
-        toRemove += id
-        boxSegmentCount -= 1
-      }
-    }while(boxCount > 0 && abs(boxes.last) > boxTarget)
-
-    if(txCount == 0 && boxCount == 0)
-      toRemove += this.id // address is empty after rollback, delete
-    else
-      toSave += this // save the changes made to this address
-
-    // Save changes
-    _history.historyStorage.insertExtra(Array.empty, toSave.toArray)
-
-    toRemove.toArray
+   * Add transaction index
+   *
+   * @param tx - numeric transaction index
+   * @return this
+   */
+  override private[extra] def addTx(tx: Long): IndexedErgoAddress = {
+    if (txs.lastOption.getOrElse(-1) != tx) txs += tx // check for duplicates
+    this
   }
 
-  /**
-    * Create an array addresses each containing a "segmentTreshold" number of this address's transaction and box indexes.
-    * These special addresses have their ids calculated by "txSegmentId" and "boxSegmentId" respectively.
-    * @return array of addresses
-    */
-  private[extra] def splitToSegments(implicit segmentTreshold: Int): Array[IndexedErgoAddress] = {
-    val data: Array[IndexedErgoAddress] = new Array[IndexedErgoAddress]((txs.length / segmentTreshold) + (boxes.length / segmentTreshold))
-    var i: Int = 0
-
-    // Split txs until under segmentTreshold
-    while(txs.length >= segmentTreshold) {
-      data(i) = new IndexedErgoAddress(txSegmentId(treeHash, txSegmentCount), txs.take(segmentTreshold), ArrayBuffer.empty[Long], None)
-      i += 1
-      txSegmentCount += 1
-      txs.remove(0, segmentTreshold)
-    }
-
-    // Split boxes until under segmentTreshold
-    while(boxes.length >= segmentTreshold) {
-      data(i) = new IndexedErgoAddress(boxSegmentId(treeHash, boxSegmentCount), ArrayBuffer.empty[Long], boxes.take(segmentTreshold), None)
-      i += 1
-      boxSegmentCount += 1
-      boxes.remove(0, segmentTreshold)
-    }
-    data
-  }
 }
 
 object IndexedErgoAddressSerializer extends ErgoSerializer[IndexedErgoAddress] {
@@ -302,45 +78,17 @@ object IndexedErgoAddressSerializer extends ErgoSerializer[IndexedErgoAddress] {
     */
   def hashErgoTree(tree: ErgoTree): ModifierId = bytesToId(Algos.hash(tree.bytes))
 
-  /**
-    * Calculates id of an address segment containing box indexes.
-    * @param hash       - hash of the parent addresses ErgoTree
-    * @param segmentNum - numberic identifier of the segment
-    * @return calculated ModifierId
-    */
-  def boxSegmentId(hash: ModifierId, segmentNum: Int): ModifierId = bytesToId(Algos.hash(hash + " box segment " + segmentNum))
-
-  /**
-    * Calculates id of an address segment transaction indexes.
-    * @param hash       - hash of the parent addresses ErgoTree
-    * @param segmentNum - numberic identifier of the segment
-    * @return calculated ModifierId
-    */
-  def txSegmentId(hash: ModifierId, segmentNum: Int): ModifierId = bytesToId(Algos.hash(hash + " tx segment " + segmentNum))
-
   override def serialize(iEa: IndexedErgoAddress, w: Writer): Unit = {
-    w.putBytes(fastIdToBytes(iEa.treeHash))
-    w.putUInt(iEa.txs.length)
-    cfor(0)(_ < iEa.txs.length, _ + 1) { i => w.putLong(iEa.txs(i))}
-    w.putUInt(iEa.boxes.length)
-    cfor(0)(_ < iEa.boxes.length, _ + 1) { i => w.putLong(iEa.boxes(i))}
+    w.putBytes(iEa.serializedId)
     w.putOption[BalanceInfo](iEa.balanceInfo)((ww, bI) => BalanceInfoSerializer.serialize(bI, ww))
-    w.putInt(iEa.boxSegmentCount)
-    w.putInt(iEa.txSegmentCount)
+    SegmentSerializer.serialize(iEa, w)
   }
 
   override def parse(r: Reader): IndexedErgoAddress = {
     val addressHash: ModifierId = bytesToId(r.getBytes(32))
-    val txnsLen: Long = r.getUInt()
-    val txns: ArrayBuffer[Long] = ArrayBuffer.empty[Long]
-    cfor(0)(_ < txnsLen, _ + 1) { _ => txns += r.getLong()}
-    val boxesLen: Long = r.getUInt()
-    val boxes: ArrayBuffer[Long] = ArrayBuffer.empty[Long]
-    cfor(0)(_ < boxesLen, _ + 1) { _ => boxes += r.getLong()}
-    val balanceInfo: Option[BalanceInfo] = r.getOption[BalanceInfo](BalanceInfoSerializer.parse(r))
-    val iEa: IndexedErgoAddress = new IndexedErgoAddress(addressHash, txns, boxes, balanceInfo)
-    iEa.boxSegmentCount = r.getInt()
-    iEa.txSegmentCount = r.getInt()
+    val iEa: IndexedErgoAddress = new IndexedErgoAddress(addressHash)
+    iEa.balanceInfo = r.getOption[BalanceInfo](BalanceInfoSerializer.parse(r))
+    SegmentSerializer.parse(r, iEa)
     iEa
   }
 }
@@ -349,81 +97,4 @@ object IndexedErgoAddress {
 
   val extraIndexTypeId: ExtraIndexTypeId = 15.toByte
 
-  /**
-    * Calculate the segment offsets for the given range.
-    * @param offset - items to skip from the start
-    * @param limit  - items to retrieve
-    * @return array of offsets
-    */
-  private def getSegmentsForRange(offset: Int, limit: Int)(implicit segmentTreshold: Int): Array[Int] =
-    (math.max(math.floor(offset * 1F / segmentTreshold).toInt, 1) to math.ceil((offset + limit) * 1F / segmentTreshold).toInt).toArray
-
-  /**
-    * Get a range of elements from an ArrayBuffer by removing the last "offset" elements,
-    * then getting the last "limit" elements reversed.
-    * @param arr    - array to get range from
-    * @param offset - number of items to skip from the end
-    * @param limit  - number of items to retrieve
-    * @return a reversed range in "arr" ArrayBuffer
-    */
-  private def sliceReversed(arr: ArrayBuffer[Long], offset: Int, limit: Int): ArrayBuffer[Long] =
-    arr.slice(arr.length - limit - offset, arr.length - offset).reverse
-
-  /**
-    * Get an array of transactions with full bodies from an array of numeric transaction indexes
-    * @param arr     - array of numeric transaction indexes to retrieve
-    * @param history - database handle
-    * @return array of transactions with full bodies
-    */
-  private def getTxs(arr: ArrayBuffer[Long], history: ErgoHistoryReader): Array[IndexedErgoTransaction] = // sorted to match explorer
-    arr.map(n => NumericTxIndex.getTxByNumber(history, n).get.retrieveBody(history)).toArray.sortBy(tx => (-tx.height, tx.id))
-
-  /**
-    * Get an array of boxes from an array of numeric box indexes
-    * @param arr     - array of numeric box indexes to retrieve
-    * @param history - database handle
-    * @return array of boxes
-    */
-  private def getBoxes(arr: ArrayBuffer[Long], history: ErgoHistoryReader): Array[IndexedErgoBox] =
-    arr.map(n => NumericBoxIndex.getBoxByNumber(history, n).get).toArray
-
-  /**
-    * Get a set of address segments from database containing numeric transaction or box indexes. Then actually retreive these indexes.
-    * @param history - database handle
-    * @param treeHash - hash of parent ErgoTree
-    * @param offset - number of items to skip from the start
-    * @param limit - max number of item to be returned
-    * @param segmentCount - number of segments of the parent address
-    * @param array - the indexes already in memory
-    * @param idOf - function to calculate segment ids, either [[txSegmentId]] or [[boxSegmentId]]
-    * @param arraySelector - function to select index array from retreived segments
-    * @param retreive - function to retreive indexes from database
-    * @tparam T - type of desired indexes, either [[IndexedErgoTransaction]] or [[IndexedErgoBox]]
-    * @return
-    */
-  private def getFromSegments[T : ClassTag](history: ErgoHistoryReader,
-                                             treeHash: ModifierId,
-                                             offset: Int,
-                                             limit: Int,
-                                             segmentCount: Int,
-                                             array: ArrayBuffer[Long],
-                                             idOf: (ModifierId, Int) => ModifierId,
-                                             arraySelector: IndexedErgoAddress => ArrayBuffer[Long],
-                                             retreive: (ArrayBuffer[Long], ErgoHistoryReader) => Array[T])
-                                            (implicit segmentTreshold: Int): Array[T] = {
-    val total: Int = segmentTreshold * segmentCount + array.length
-    if(offset >= total)
-      return Array.empty[T] // return empty array if all elements are skipped
-    if (offset + limit > array.length && segmentCount > 0) {
-      val data: ArrayBuffer[Long] = ArrayBuffer.empty[Long]
-      getSegmentsForRange(offset, limit).map(n => math.max(segmentCount - n, 0)).distinct.foreach { num =>
-        arraySelector(
-          history.typedExtraIndexById[IndexedErgoAddress](idOf(treeHash, num)).get
-        ) ++=: data
-      }
-      data ++= (if (offset < array.length) array else Nil)
-      retreive(sliceReversed(data, offset % segmentTreshold, math.min(total - offset, limit)), history)
-    } else
-      retreive(sliceReversed(array, offset, limit), history)
-  }
 }
